Some peripheral support for the Daisy Seed stm32H750 board.
- Added the SDRAM controller and set up heap to use the 64 MB external RAM
- Added the rest of the board pin definitions from the pinout
- Did peripheral pins cleanup for the stm32H750 chip
- Added SDMMC peripheral and sdioio module support
- Added pwmio support and started work on audiopwmio

+++ ports/stm/peripherals/timers.h
@@ -7,13 +7,12 @@
 #pragma once
 
 #include <stdint.h>
-#include "py/mphal.h"
-#include "peripherals/periph.h"
+#include <stdbool.h>
 
 #include STM32_HAL_H
 
-void tim_clock_enable(uint16_t mask);
-void tim_clock_disable(uint16_t mask);
+void tim_clock_enable(uint32_t mask);
+void tim_clock_disable(uint32_t mask);
 uint32_t stm_peripherals_timer_get_source_freq(TIM_TypeDef *timer);
 size_t stm_peripherals_timer_get_irqnum(TIM_TypeDef *instance);
 void timers_reset(void);
